requirements.txt
openai>=1.0.0
pathlib>=1.0.1 
ollama>=0.1.0
weaviate-client>=4.0.0 